+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -122,6 +122,11 @@ New checks
   Warns if a function name is used in an expression other than a call
   or address-of.
 
+- New :doc:`bsl-identifier-typographically-unambiguous
+  <clang-tidy/checks/bsl-identifier-typographically-unambiguous>` check.
+
+  FIXME: add release notes.
+
 - New :doc:`bsl-lambda-implicit-capture
   <clang-tidy/checks/bsl-lambda-implicit-capture>` check.